# docs/openapi/declaration.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RÉCOR Declaration Service",
    "description": "REST surface of the RÉCOR Declaration service. Accepts beneficial-ownership declarations, returns signed receipts, relays verification outcomes from the Verification Engine, and exposes operator endpoints under `/v1/internal/*`. Endpoints under the `internal` tag are not for general consumer use.",
    "contact": {
      "name": "RÉCOR domain team",
      "email": "noreply@recor.cm"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local dev"
    },
    {
      "url": "https://api.recor.cm",
      "description": "Production (TBD)"
    }
  ],
  "paths": {
...
    "/v1/declarations": {
      "post": {
        "tags": [
          "declarations"
        ],
        "operationId": "submitDeclaration",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Optional client-supplied idempotency key. Replays of the same key with the same body return the original response.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitDeclarationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Idempotent replay; returns the recorded response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitDeclarationResponse"
                }
              }
            }
          },
          "201": {
            "description": "Declaration accepted and persisted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitDeclarationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid bearer token or bad attestation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Attestation principal mismatch / authorisation denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict OR optimistic concurrency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate-limited (OPS-1; token-bucket per principal)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "devPrincipalHeader": []
          }
        ]
      }
    },
    "/v1/declarations/{declaration_id}": {
      "get": {
        "tags": [
          "declarations"
        ],
        "operationId": "getDeclaration",
        "parameters": [
          {
            "name": "declaration_id",
            "in": "path",
            "description": "Declaration UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current projection of the declaration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDeclarationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Declaration is owned by a different principal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Declaration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "devPrincipalHeader": []
          }
        ]
      }
    },
    "/v1/declarations/{declaration_id}/supersede": {
      "post": {
        "tags": [
          "declarations"
        ],
        "operationId": "supersedeDeclaration",
        "parameters": [
          {
            "name": "declaration_id",
            "in": "path",
            "description": "Identifier of the declaration to supersede",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitDeclarationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successor declaration accepted; previous record marked superseded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupersedeDeclarationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid bearer token or bad attestation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not the owner of the prior declaration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Prior declaration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Already superseded or optimistic-concurrency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate-limited (OPS-1)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "devPrincipalHeader": []
          }
        ]
      }
    },
    "/v1/internal/outbox-dlq": {
      "get": {
        "tags": [
          "internal"
        ],
        "operationId": "listDlq",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of rows to return; defaults to 50.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 50
            },
            "example": 50
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Zero-indexed offset for paging.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0
            },
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Page of DLQ rows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDlqResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Principal not on the admin allowlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Backend failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Admin endpoints disabled (allowlist empty)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "devPrincipalHeader": []
          }
        ]
      }
    },
    "/v1/internal/outbox-dlq/{id}/replay": {
      "post": {
        "tags": [
          "internal"
        ],
        "operationId": "replayDlq",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "DLQ row id to replay",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Row was moved back to the outbox for retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayDlqResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Principal not on the admin allowlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No DLQ row with the given id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Backend failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Admin endpoints disabled (allowlist empty)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "devPrincipalHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "description": "Standard error response body. Every non-2xx response is shaped like\n`{ \"error\": { \"kind\": \"<machine>\", \"message\": \"<human>\" } }`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "ListDlqResponse": {
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DlqItem"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int64",
            "description": "Limit echoed back from the query string."
          },
          "offset": {
            "type": "integer",
            "format": "int64",
            "description": "Offset echoed back from the query string."
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of DLQ rows currently held (unfiltered)."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OIDC-issued JWT. The service verifies the signature against the configured issuer's JWKS, plus `iss`/`aud`/`exp`/`nbf` claims. HMAC algorithms are refused (algorithm-confusion mitigation)."
      },
      "devPrincipalHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Recor-Dev-Principal",
        "description": "Dev-only principal shortcut. Refused outside ENVIRONMENT=dev. Production must always use `bearerAuth`."
      },
      "hmacSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RECOR-Signature",
        "description": "HMAC-SHA256 over the raw request body, hex-encoded, using the shared writeback secret. Constant-time verification; rotation supported via a transient old-secret window."
      }
    }
  },
...
}